+++ packages/plugin-workflow-approval/src/client/configuration/trigger-approval/initializers/ApplyFormDetail.tsx
@@ -9,8 +9,8 @@ import {
 } from '@tachybase/client';
 
 import { APPROVAL_INITIATION_STATUS } from '../../../common/constants/approval-initiation-status';
+import { flatSchemaArray } from '../../../common/tools/flatSchemaArray';
 import { NAMESPACE } from '../../../locale';
-import { flatSchemaArray } from '../../../usage/pc/constants';
 
 // 添加卡片-Initializer的component
 export const ApplyFormDetailInitializerItem = () => {
